BooleanCondition -> BooleanFunction
